python_job/requirements.txt
Django==2.2.2
django-ckeditor==5.7.1
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-filter==2.1.0
django-js-asset==1.2.2
django-taggit==1.1.0
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
